requirements.txt
Flask==2.0.2
python-can==3.3.4
pynmea2==1.15.0
gunicorn==20.1.0
Werkzeug==2.0.3
pyserial==3.5
